# backend/swagger.yaml
swagger: "2.0"
info:
  description: "A web-based  judging  application  that  allows  users,  namely Admin, to host events and Participants, to create their project ideas or join an existing project and Judges, to be seamlessly involved in rating the assigned projects. Participants create and join the projects in an event. Eventually, judges are allowed to rate and admins can announce the winners of the projects in each event that are assigned to them. Each user has their roles and access to perform their respective operations described in the upcoming sections."
  version: "1.0.0"
  title: "SeamlessJudging"
host: "stark.cse.buffalo:3000"
schemes:
- "http"
paths:
...
  /api/v1/judge:
    get:
      tags:
      - "Judge"
      summary: "Retrieve Judges Details"
      description: "Only admin retreive judges complete list"
      operationId: "getJudges"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      responses:
          "200":
            description: "successful"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Judge details retrieved successfully"
                response_data:
                  $ref: "#/definitions/JudgeListResponse"
          "404":
            description: "Not Found"

    post:
      tags:
      - "Judge"
      summary: "Add a Judge"
      description: "Only admin can add a judge"
      operationId: "addJudge"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request to add a judge"
        required: true
        schema:
          $ref: "#/definitions/AddJudgeRequest"
      responses:
          "201":
            description: "successfully added"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Judge added succesfully"
                response_data:
                  $ref: "#/definitions/ResponseData"
          "404":
            description: "Not Found"
    
    delete:
      tags:
      - "Judge"
      summary: "Delete judges in bulk"
      description: "Delete judges in bulk"
      operationId: "deleteJudges"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request to delete judges"
        required: true
        schema:
          type: "object"
          properties:
            ids:
              $ref: "#/definitions/ArrayOfInt"
              
      responses:
          "201":
            description: "Judges deleted successfully"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Judges deleted successfully"
                response_data:
                  type: "object"
          "404":
            description: "Not Found"
  /api/v1/judge/{judgeId}:
    put:
      tags:
      - "Judge"
      summary: "Update a Judge detail"
      description: "Only admin can update a judge"
      operationId: "updateJudge"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - name: "judgeId"
        in: "path"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request to update a judge"
        required: true
        schema:
          $ref: "#/definitions/AddJudgeRequest"
      responses:
          "200":
            description: "successfully updated"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Judge updated succesfully"
                response_data:
                  $ref: "#/definitions/ResponseData"
          "404":
            description: "Not Found"
...
definitions:
...
  ResponseData:
    type: object
    properties:
      id:
        type: integer
        example: 1
  ArrayOfInt:
    type: array
    items:
      type: integer
...
  JudgeListResponse:
    type: "object"
    properties:
      user_id:
          type: integer
          example: 1
      first_name:
          type: "string"
          example: "John"
      middle_name:
          type: "string"
          example: ""
      last_name:
          type: "string"
          example: "Doe"
      email:
          type: "string"
          example: "john@abc.com"
      image:
          type: "string"
          example: "C://logo.png"
      status:
          type: "string"
          example: "ACTIVE"
      events:
        type: "array"
        items:
          type: "object"
          properties:
            event_id:
              type: "integer"
              example: 1
            name:
              type: "string"
              example: "Summer 2022"
            location:
              type: "string"
              example: "Davis 301"
            logo:
              type: "string"
              example: "C://event_logo.jpg"
            date:
              type: "string"
              example: ["2022-06-06T01:43:43.000Z", "2022-07-06T04:43:43.000Z"]
            code:
              type: "string"
              example: "12345432"
  AddJudgeRequest:
    type: "object"
    properties:
      first_name:
          type: "string"
          example: "John"
      middle_name:
          type: "string"
          example: ""
      last_name:
          type: "string"
          example: "Doe"
      email:
          type: "string"
          example: "john@abc.com"